requirements.txt
Pillow
Django
djangorestframework
django-cors-headers